Added /api/replicas/tokens/ endpoint
This endpoint returns name, apikey for each active replica, but only if
the authenticated user is an administrator. Still to do is add user
based authentication for this check so that the data can be fetched via
geonet on the command line. At least for now, the JSON can be copied and
pasted easily.

Fixes #7