Better authentication and machine permissions
Implemented a subclass of TokenAuthentication that allows both users and
machine devices to authenticate with a token or api key. Changes the
headers to more traditional authorization format.

Implemented machine user and active machine user permissions so that
inactive hosts can heartbeat but can't post to latency endpoints.

Fixes #15
Fixes #16