Enhanced replicas endpoint with modifying methods
Refactored the ReplicaViewSet to a ModelViewSet to give it detail,
create, update, and delete endpoints. Modified the serializer to include
more data (breaks peers/kekahu sync) but is necessary to allow for
replica management on the API. Also added an IsAdminOrReadOnly
permission so that only admins can modify replicas, but everyone else
can read.

Added an activate action that takes a flag as an argument and can
activate or deactivate replicas. Added a button to the replica detail
that PUTS an activate or deactivate request to toggle the status of the
replica.

Fixes #18
Fixes #19